feat 자동차 경주기록, 우승자 그려주는 객체
