--- src/main/java/carRace/step5/domain/util/CarRasingDraw.java
@@ -1,14 +1,0 @@
-package carRace.step5.domain.util;
-
-public class CarRasingDraw {
-
-    public static String changeNumberToChar(int count) {
-        StringBuilder sb = new StringBuilder();
-
-        for (int i = 0; i < count; i++) {
-            sb.append("-");
-        }
-
-        return sb.toString();
-    }
-}
